+++ lib/internal/crypto/x509.js
@@ -371,7 +371,7 @@ class X509Certificate {
     // TODO(tniessen): do not depend on translatePeerCertificate here, return
     // the correct legacy representation from the binding
     lazyTranslatePeerCertificate ??=
-      require('_tls_common').translatePeerCertificate;
+      require('internal/tls/common').translatePeerCertificate;
     return lazyTranslatePeerCertificate(this[kHandle].toLegacy());
   }
 }
