+++ lib/tls.js
@@ -55,8 +55,8 @@ const {
 } = internalBinding('crypto');
 const { Buffer } = require('buffer');
 const { canonicalizeIP } = internalBinding('cares_wrap');
-const _tls_common = require('_tls_common');
-const _tls_wrap = require('_tls_wrap');
+const tlsCommon = require('internal/tls/common');
+const tlsWrap = require('internal/tls/wrap');
 const { validateString } = require('internal/validators');
 
 // Allow {CLIENT_RENEG_LIMIT} client-initiated session renegotiations
@@ -386,9 +386,9 @@ exports.checkServerIdentity = function checkServerIdentity(hostname, cert) {
   }
 };
 
-exports.createSecureContext = _tls_common.createSecureContext;
-exports.SecureContext = _tls_common.SecureContext;
-exports.TLSSocket = _tls_wrap.TLSSocket;
-exports.Server = _tls_wrap.Server;
-exports.createServer = _tls_wrap.createServer;
-exports.connect = _tls_wrap.connect;
+exports.createSecureContext = tlsCommon.createSecureContext;
+exports.SecureContext = tlsCommon.SecureContext;
+exports.TLSSocket = tlsWrap.TLSSocket;
+exports.Server = tlsWrap.Server;
+exports.createServer = tlsWrap.createServer;
+exports.connect = tlsWrap.connect;
